% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_fertility_rates.R
\name{construct_fertility_rates}
\alias{construct_fertility_rates}
\title{calcualtes age specific and total fertility rates per year}
\usage{
construct_fertility_rates(data, pops)
}
\arguments{
\item{data}{births data frame}

\item{pops}{population data frame}
}
\value{
tablulated age specific and total fertility rates per year
}
\description{
calcualtes age specific and total fertility rates per year
}
\examples{
fertility_rates <- construct_fertility_rates(bth_data, population)

}
