% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_age.R
\name{construct_age}
\alias{construct_age}
\title{Construct and add Age Column to Dataset}
\usage{
construct_age(dataset, dob_col, age_col = "birth3b", end_date_col = NULL)
}
\arguments{
\item{dataset}{A data frame containing at least one column with date of birth values.}

\item{dob_col}{The name of the column containing date of birth values.}

\item{age_col}{The name of the new column to store the calculated age. Default is `"age"`.}

\item{end_date_col}{Optional. The name of the column containing end date values. If not provided, today's date is used.}
}
\value{
A data frame with an additional column containing the calculated age.
}
\description{
This function takes a dataset with a specified date of birth column and returns
the dataset with an additional column which contains the calculated age.
}
\examples{
df <- data.frame(death2a = as.Date(c("1990-05-15", "1985-10-30", "2000-07-22")),
                 death1a = as.Date(c("2022-07-15", "2022-07-15", "2022-07-15")))
new_df <- construct_age(df, dob_col = "death2a", age_col = "death2b", end_date_col = "death1a")
print(new_df)

df <- data.frame(birth3a = as.Date(c("1990-05-15", "1985-10-30", "2000-07-22")))
new_df <- construct_age(df, dob_col = "birth3a")
print(new_df)
}
