% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_age_group.R
\name{construct_age_group}
\alias{construct_age_group}
\title{Construct Age Group Column to Dataset}
\usage{
construct_age_group(dataset, age_col, group_size = 5, new_col = "fert_age_grp")
}
\arguments{
\item{dataset}{A data frame containing at least one column with age values.}

\item{age_col}{The name of the column containing age values.}

\item{group_size}{An integer specifying the size of the age groups. Default is 5.}

\item{new_col}{The name of the new column to store the age groups. Default is "fert_age_grp".}
}
\value{
A data frame with the new column added, containing age groups.
}
\description{
This function takes a dataset and adds a new column that contains age groups.
}
\examples{
df <- data.frame(id = 1:10, age = c(23, 45, 34, 56, 29, 12, 67, 44, 33, 55))
new_df <- construct_age_group(df, age_col = "age", group_size = 10)
print(new_df)
}
