% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_cause_flag.R
\name{construct_cause_flag}
\alias{construct_cause_flag}
\title{construct_cause_flag}
\usage{
construct_cause_flag(dataset, code_pattern, column_pattern, new_column_name)
}
\arguments{
\item{dataset}{The name of the dataframe being used.}

\item{code_pattern}{The code, or codes, you are searching for.}

\item{column_pattern}{The name, or partial name, of the column(s) being searched.}

\item{new_column_name}{The name the column will be called.}
}
\value{
A new column containing a 1 or 0 to flag the record.
}
\description{
Creates a flag variable based on supplied ICD codes.
}
\examples{
data <- data \%>\% construct_cause_flag("E85|E619", "FIC10MEN", "CHECK")

}
