% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_empty_var.R
\name{construct_empty_var}
\alias{construct_empty_var}
\title{Add NA Column to Dataset}
\usage{
construct_empty_var(dataset, column_name = "birth1j")
}
\arguments{
\item{dataset}{A data frame to which the new column will be added.}

\item{column_name}{The name of the new column to be added. Default is `"birth1j"`.}
}
\value{
A data frame with the new column added, containing NA values.
}
\description{
This function takes a dataset and adds a new column filled with NA values.
}
\examples{
df <- data.frame(id = 1:5, name = c("Alice", "Bob", "Charlie", "David", "Eve"))
new_df <- construct_empty_var(df)
print(new_df)
}
