% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_leading_groups.R
\name{construct_leading_groups}
\alias{construct_leading_groups}
\title{Create Leading deaths Age Group Column}
\usage{
construct_leading_groups(
  data,
  age_col = "death2b",
  age_group_col = "age_grp_lead"
)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{age_col}{The column name (as a string) containing age in integer form.}

\item{age_group_col}{The new column name for the age group labels (default is "age_grp_lead").}
}
\value{
A data frame with the new age group column added.
}
\description{
This function takes a dataset and a column containing age in integer form,
and creates a new column with age group labels.
}
\examples{
data <- data.frame(death2b = c(3, 8, 45, 72))
data <- construct_leading_groups(data, age_col = 'death2b', age_group_col = "age_grp_lead")
print(data)
}
