% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_sample_data.R
\name{construct_sample_death_data}
\alias{construct_sample_death_data}
\title{Read Sample Death Data}
\usage{
construct_sample_death_data()
}
\value{
A data frame containing the sample death data.
}
\description{
Reads the sample data file created_death_data.csv from the package's extdata directory.
}
