% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_sample_data.R
\name{construct_sample_divorce_data}
\alias{construct_sample_divorce_data}
\title{Read Sample Divorce Data}
\usage{
construct_sample_divorce_data()
}
\value{
A data frame containing the sample divorce data.
}
\description{
Reads the sample data file created_divorce_data.csv from the package's extdata directory.
}
