% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_sample_data.R
\name{construct_sample_estim_death_data}
\alias{construct_sample_estim_death_data}
\title{Read Sample Estimated Death Data (i.e. "expected" deaths by region, age, sex)}
\usage{
construct_sample_estim_death_data()
}
\value{
A data frame containing the sample death data.
}
\description{
Reads the sample data file created_death_estim.csv from the package's extdata directory.
}
