% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_sample_data.R
\name{construct_sample_pop_estim}
\alias{construct_sample_pop_estim}
\title{Read Sample Population Estimates Data}
\usage{
construct_sample_pop_estim()
}
\value{
A data frame containing the sample marriage data.
}
\description{
Reads the sample data file created_population_estim.csv from the package's extdata directory.
}
