% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_year.R
\name{construct_year}
\alias{construct_year}
\title{Add Year Column to Dataset}
\usage{
construct_year(dataset, date_col = "birth1c", year_col = "dobyr")
}
\arguments{
\item{dataset}{A data frame containing at least one column with date values.}

\item{date_col}{The name of the column containing date values. Default is `"birth1c"`.}

\item{year_col}{The name of the column containing year values. Default is `"dobyr"`.}
}
\value{
A data frame with an additional column `dobyr` containing the year of the date column.
}
\description{
This function takes a dataset with a specified date column and returns
the dataset with an additional column `dobyr` which contains the year extracted from the date column.
}
\examples{
df <- data.frame(birth1c = c("1990-05-15", "1985-10-30", "2000-07-22"))
new_df <- add_year_column(df)
print(new_df)

df <- data.frame(event_date = c("1990-05-15", "1985-10-30", "2000-07-22"))
new_df <- add_year_column(df, date_col = "event_date")
print(new_df)
}
