% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_config_args.R
\name{convert_config_args}
\alias{convert_config_args}
\title{convert_config_args}
\usage{
convert_config_args(args)
}
\arguments{
\item{args}{A list of variables that comes from the config file.}
}
\value{
A list of arguments that can be passed into the functions.
}
\description{
A function that takes the user defined arguments for the table functions and converts the format to include when calling the function.
}
\details{
Given the config defined arguments for a function, it will look for the data fields and convert them to create the function call.
}
\examples{
convert_config_args(args)

}
