% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_3_10.R
\name{create_t3.10}
\alias{create_t3.10}
\title{Create Table 3.10}
\usage{
create_t3.10(
  data,
  dth_est,
  date_var,
  data_year = NA,
  tablename = "Table_3_10",
  output_path = NULL
)
}
\arguments{
\item{data}{dataframe being used}

\item{dth_est}{dataframe being used with estimated Death data}

\item{date_var}{occurrence data being used e.g. dobyr, dodyr etc}

\item{data_year}{The year to report on. Defaults to the last in the data.}

\item{tablename}{Name of the table to be saved as a csv file. Optional.}

\item{output_path}{The path to export the generated csv table. Optional.}
}
\value{
data frame of tabulated results
}
\description{
Table 3.10 Adjustment of deaths by age group and birth2a of decedent
}
\examples{
t3.10 <- create_t3.10(dth_data, date_var = dodyr, data_year = 2022, tablename = "Table_3_10")

}
