% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_3_4_&_3_6.R
\name{create_t3.4_and_3.6}
\alias{create_t3.4_and_3.6}
\title{Creates Table 3.4 and 3.6}
\usage{
create_t3.4_and_3.6(
  data,
  est_data,
  topic = NA,
  date_var = "dobyr",
  data_year = NA,
  tablename = "Table_3_4",
  output_path = NULL
)
}
\arguments{
\item{data}{data frame being used}

\item{est_data}{data frame of estimated data being used}

\item{topic}{whether the table is births or deaths data}

\item{data_year}{The year to report on. Defaults to the last in the data.}

\item{tablename}{Name of the table to be saved as a csv file. Optional.}

\item{output_path}{The path to export the generated csv table. Optional.}

\item{by_var}{the occurrence year being used e.g. dobyr or dodyr}
}
\value{
data frames for tabulated versions of Table 3.4 and 3.6
}
\description{
Creates Table 3.4 and 3.6
}
\examples{
t3.4 <- create_t3.4_to_3.7(bth_data, bth_est, dobyr, topic = "births", tablename = "Table_3_4")
}
