% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_3_8_&_3_9.R
\name{create_t3.8_and_t3.9}
\alias{create_t3.8_and_t3.9}
\title{Creates Tables 3.8 and 3.9}
\usage{
create_t3.8_and_t3.9(
  data,
  est_data,
  date_var,
  by_var,
  topic = NA,
  data_year = NA,
  tablename = "Table_3_9",
  output_path = NULL
)
}
\arguments{
\item{data}{name of data frame being used}

\item{est_data}{data frame of estimated data being used}

\item{date_var}{occurrence data variable being used}

\item{by_var}{the occurrence year being used e.g. dobyr or dodyr}

\item{topic}{whether births or deahts is being run}

\item{data_year}{The year to report on. Defaults to the last in the data.}

\item{tablename}{Name of the table to be saved as a csv file. Optional.}

\item{output_path}{The path to export the generated csv table. Optional.}
}
\value{
data frame with tabulated results
}
\description{
Table 3.8 Adjustment of live births by place of usual residence of mother and sex of newborn

Table 3.9 Adjustment of deaths by place of usual residence and sex of decedent
}
\examples{
t3.8 <- create_t3.8_and_t3.9(data = bth_data, est_data = bth_est,
date_var = dobyr, data_year = 2022, by_var = birth1c,
topic = "births", tablename = "Table_3_8")
t3.9 <- create_t3.8_and_t3.9(data = dth_data, est_data = dth_est,
                            date_var = dodyr, data_year = 2022, by_var = birth1c,
                            topic = "deaths", tablename = "Table_3_9")

}
