% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_4_2.R
\name{create_t4.2}
\alias{create_t4.2}
\title{Creates Table 4.2}
\usage{
create_t4.2(
  data,
  est_data,
  date_var = "dobyr",
  data_year = NA,
  tablename = "Table_4_2",
  output_path = NULL
)
}
\arguments{
\item{data}{Birth records data frame}

\item{est_data}{Estimated births data frame}

\item{date_var}{Name of date variable (years)}

\item{data_year}{The year to report on. Defaults to the last in the data.}

\item{tablename}{Name of the table to be saved as a csv file. Optional.}

\item{output_path}{The path to export the generated csv table. Optional.}
}
\value{
A table of counts of births for each place of occurance split by sex.
}
\description{
Table 4.2 Live births by place of occurrence (birth1c) and sex (birth2a) of newborn
}
\examples{
t4.2 <- create_t4.2(birth_data, est_data, date_var = "dobyr",
data_year = 2023, tablename = "Table_4_2", output_path = "outputs/births/")
}
