% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_4_3.R
\name{create_t4.3}
\alias{create_t4.3}
\title{Creates Table 4.3}
\usage{
create_t4.3(
  data,
  date_var = "dobyr",
  data_year = NA,
  tablename = "Table_4_3",
  output_path = NULL
)
}
\arguments{
\item{data}{data frame being used}

\item{date_var}{occurrence data being used e.g. dobyr, dodyr etc}

\item{data_year}{The year to report on. Defaults to the last in the data.}

\item{tablename}{Name of the table to be saved as a csv file. Optional.}

\item{output_path}{The path to export the generated csv table. Optional.}
}
\value{
data frame with tablutated results
}
\description{
Table 4.3 Live births by place of occurrence and place of usual residence of mother
}
\examples{
t4.3 <- create_t4.3(bth_data, dobyr, 2022)
}
