% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_4_4_to_4_6.R
\name{create_t4.4_to_4_6}
\alias{create_t4.4_to_4_6}
\title{creates Tables 4.5, 4.5 and 4.6}
\usage{
create_t4.4_to_4_6(
  data,
  data_year = NA,
  col_var = fert_age_grp,
  by_var = birth1g,
  rural_urban = "no",
  tablename = NA
)
}
\arguments{
\item{data}{the dataframe being used}

\item{data_year}{the year of the data needed}

\item{col_var}{variable of data being used e.g. fert_age_grp}

\item{by_var}{the variable the data is being grouped by}

\item{rural_urban}{if this is no then no rural/urban split is used. Else specify "rural" or "urban"}

\item{tablename}{name of the table being saved as a csv file}
}
\value{
tabulated data
}
\description{
Table 4.4 Live births by age of mother and type of birth

Table 4.5 Live births by age and marital status of mother, urban areas

Table 4.6 Live births by age and marital status of mother, rural areas
}
\examples{
t4.4 <- create_t4.4_to_4_6(bth_data, data_year = 2022, col_var = fert_age_grp, by_var = birth1g, rural_urban = "no", tablename = "Table_4_4")
t4.5 <- create_t4.4_to_4_6(bth_data, year = 2022, col_var = fert_age_grp, by_var = birth3c, rural_urban = "no", tablename = "Table_4_5")
t4.6 <- create_t4.4_to_4_6(bth_data, year = 2022, col_var = fert_age_grp, by_var = birth3c, rural_urban = "no", tablename = "Table_4_6")
}
