% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_4_7.R
\name{create_t4.7}
\alias{create_t4.7}
\title{Create Table 4.7}
\usage{
create_t4.7(data, date_var, data_year = NA, tablename = NA)
}
\arguments{
\item{data}{data frame being used}

\item{date_var}{occurrence data variable being used}

\item{data_year}{year of data}

\item{tablename}{name of the table being saved as a csv file}
}
\value{
data frames for tabulated versions of Table 4.7
}
\description{
Table 4.7 Live births by place of occurrence, site of delivery and attendant at birth
}
\examples{
t4.7 <- create_table_4.7(bth_data, dobyr, 2022, tablename = "Table_4_7")

}
