% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_4_8.R
\name{create_t4.8}
\alias{create_t4.8}
\title{creates table 4.8}
\usage{
create_t4.8(
  data,
  est_data,
  pops,
  date_var,
  data_year = NA,
  by_var = NA,
  tablename = "Table_4_8",
  output_path = NULL
)
}
\arguments{
\item{data}{data frame being used}

\item{est_data}{data frame for estimate data}

\item{pops}{estimated population sizes by year}

\item{date_var}{occurrence data variable being used}

\item{data_year}{year of data}

\item{by_var}{variable the data is grouped by}

\item{tablename}{name for csv output use _ instead of . for names}

\item{output_path}{The path to export the generated table}
}
\value{
data frame with tabulated result
}
\description{
Table 4.8 Crude birth rate (CBR) by place of usual residence of mother
}
\examples{
t4.8 <- create_t4.8(bth_data, bth_est, dobyr, data_year = 2022, by_var = birth1c, tablename = "Table_4_8")
}
