% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_5_3_&_5_4.R
\name{create_t5.3_and_t5.4}
\alias{create_t5.3_and_t5.4}
\title{Creates Tables 5.3 & 5.4}
\usage{
create_t5.3_and_t5.4(
  data,
  sex_filter = "male",
  date_var = "dodyr",
  data_year = NA,
  tablename = "Table_5_3",
  output_path = NULL
)
}
\arguments{
\item{data}{Death records data frame.}

\item{sex_filter}{Sex to filter by. Defaults to "male".}

\item{date_var}{Event year column of the dataframe. Defaults to "dodyr".}

\item{data_year}{The year to report on. Defaults to the last in the data.}

\item{tablename}{Name of the table to be saved as a csv file. Optional.}

\item{output_path}{The path to export the generated csv table. Optional.}
}
\value{
A data frame of tabulated results
}
\description{
Table 5.3 Deaths by place of occurrence and place of usual residence of decedent (males)

Table 5.4 Deaths by place of occurrence and place of usual residence of decedent (females)

Variables needed per record:

- `dodyr`: Year of death.

- `death1c`: Region of death.

- `death2o`: Region of usual residence.

- `death2c`: Sex of the deceased.
}
\examples{
death_data <- construct_sample_death_data()
death_data <- construct_year(death_data, date_col = "death1a", year_col = "dodyr")
table_5_3 <- create_t5.3_and_t5.4(death_data, data_year = 2022)
}
