% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_6_1.R
\name{create_t6.1}
\alias{create_t6.1}
\title{Create Table 6.1}
\usage{
create_t6.1(data, cause, date_var, data_year = NA)
}
\arguments{
\item{data}{data frame being used}

\item{cause}{cause of death data frame}

\item{date_var}{column containing the year of event occurence}

\item{data_year}{year of data}
}
\value{
Returns data frame with tabulated results
}
\description{
Table 6.1 Leading causes of death by broad age group and sex
}
\examples{
t6.1 <- create_t6.1(data = dth_data, date_var = dodyr, data_year = 2022)

}
