% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_6_5_to_6_10.R
\name{create_t6.5_t6.10}
\alias{create_t6.5_t6.10}
\title{Creates Tables 6.5 to 6.10}
\usage{
create_t6.5_t6.10(data, data_year = NA, sex_value = "male", age_group = "<5")
}
\arguments{
\item{data}{data frame being used}

\item{sex_value}{whether male or female}

\item{age_group}{which age group is being used e.g. "<5"}
}
\value{
data frame with tabulated results
}
\description{
Table 6.5 Ten leading causes of death, infants and children (0–4 years, both sexes combined)

Table 6.6 Ten leading causes of death, children (5–14 years, both sexes combined)

Table 6.7 Ten leading causes of death, adolescents and adults (15–69 years, males)

Table 6.8 Ten leading causes of death, adolescents and adults (15–69 years, females)

Table 6.9 Ten leading causes of death, older adults (70+ years, males)

Table 6.10 Ten leading causes of death, older adults (70+ years, females),
}
\examples{
t6.5 <- create_t6.5_t6.10(dth_data, sex_value = c("male","female"), age_group = "<5")
t6.6 <- create_t6.5_t6.10(dth_data, sex_value = c("male","female"), age_group = "5-14")

}
