% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_7_1.R
\name{create_t7.1}
\alias{create_t7.1}
\title{Generate Table 7.1 Summary statistics on marriages and divorces by year of occurrence}
\usage{
create_t7.1(
  marriage_data,
  divorce_data,
  population_data,
  date_var = "domyr",
  data_year = NA
)
}
\arguments{
\item{marriage_data}{The marriage data.}

\item{divorce_data}{The divorce data.}

\item{population_data}{The population data.}

\item{date_var}{The variable to use the year from}

\item{data_year}{A particular year to choose from}
}
\value{
A data frame containing the summary statistics on marriages and divorces by year.
}
\description{
This function takes marriage and divorce data and generates a table with summary statistics on marriages and divorces by year.
}
\examples{
create_t7.1(marriage_data, divorce_data, population_data)
}
