% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_8_1.R
\name{create_t8.1}
\alias{create_t8.1}
\title{Creat Table 8.1}
\usage{
create_t8.1(tablename = NA)
}
\arguments{
\item{tablename}{name for csv output use _ instead of . for names}
}
\value{
data frame with tabulated result
}
\description{
Table 8.1 Live births, deaths, and infant and child deaths by year of occurrence
}
\examples{
t8.1 <- create_t8.1(tablename = "Table _8_1")
}
