% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_7_4_&_7_5.R
\name{create_table_7.4_and_7.5}
\alias{create_table_7.4_and_7.5}
\title{Create Tables 7.4 and 7.5}
\usage{
create_table_7.4_and_7.5(
  data,
  data_year = NA,
  groombride = "groom",
  tablename = "Table_7_4",
  output_path = NULL
)
}
\arguments{
\item{data}{dataframe being used}

\item{data_year}{The year to report on. Defaults to the last in the data.}

\item{groombride}{whether the table is for groom or bride data.}

\item{tablename}{Name of the table to be saved as a csv file. Optional.}

\item{output_path}{The path to export the generated csv table. Optional.}
}
\value{
data frame of tabulated results
}
\description{
Table 7.4 Marriages by age of groom and previous marital status

Table 7.5 Marriages by age of bride and previous marital status
}
\examples{
t7.4 <- create_table_7.4_and_7.5(marr_data, data_year = 2020, groombride = "groom", tablename = "Table_7_4")
t7.5 <- create_table_7.4_and_7.5(marr_data, data_year = 2020, groombride = "bride", tablename = "Table_7_5")

}
