% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_excel.R
\name{csv_to_excel}
\alias{csv_to_excel}
\title{Convert CSV files into an Excel spreadsheet}
\usage{
csv_to_excel(path = ".")
}
\arguments{
\item{path}{The path to the folder containing the CSV files}
}
\value{
An Excel file (.xlsx) containing worksheets for each imported CSV file
}
\description{
Reads multiple CSV files and outputs them as individual worksheets within a single Excel spreadsheet.
}
\examples{
csv_to_excel(path = ".")

}
