% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_cause.R
\name{flag_cause}
\alias{flag_cause}
\title{flag_cause}
\usage{
flag_cause(dataset, code_pattern, column_pattern, new_column_name)
}
\arguments{
\item{dataset}{The name of the dataframe being used.}

\item{code_pattern}{The code, or codes, you are searching for.}

\item{column_pattern}{The name, or partial name, of the column(s) being searched.}

\item{new_column_name}{The name the column will be called.}
}
\value{
A new column containing a 1 or 0 to flag the record.
}
\description{
Creates a flag variable based on supplied ICD codes.
}
\examples{
data <- data \%>\% flag_cause("E85|E619", "FIC10MEN", "CHECK")

}
