% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_csv_xlsx.R
\name{handle_csv_xlsx}
\alias{handle_csv_xlsx}
\title{Convert CSV Files to an Excel Workbook}
\usage{
handle_csv_xlsx(input_path = ".", output_path = "output.xlsx")
}
\arguments{
\item{output_path}{A character string specifying the directory and file name to write the xlsx to.}

\item{path}{A character string specifying the directory containing the CSV files. Defaults to the current working directory (".").}
}
\value{
An Excel file named "output.xlsx" containing the contents of the CSV files.
}
\description{
This function reads all CSV files in a specified directory and writes their contents to separate sheets in a single Excel workbook.
}
\examples{
# Convert CSV files in the current directory to an Excel workbook
handle_csv_xlsx(path = ".")

# Convert CSV files in a specified directory to an Excel workbook
handle_csv_xlsx(input_path = "path/to/directory", output_path = "path/with/file/output.xlsx")
}
