% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_table_selection.R
\name{handle_table_selection}
\alias{handle_table_selection}
\title{Handle Table Selection}
\usage{
handle_table_selection(section_num, all_tables)
}
\arguments{
\item{section_num}{A string}

\item{all_tables}{A dictionary with all the tables and their information}
}
\value{
A list of table ids that belong to the section_num number
}
\description{
A function that takes the user defined selection of tables to run from the config and will generate a list of tables to run.
}
\details{
Given a section_num number, it will go through the dictionary of all tables and return those that belong to that section by looking at the start of the name.
}
\examples{
sect_tables("3", all_tables)

}
