% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_line_chart.R
\name{plot_line_chart}
\alias{plot_line_chart}
\title{Plot Line chart}
\usage{
plot_line_chart(
  data,
  x,
  y,
  group,
  line_colour = "ByGroup",
  line_width = 1,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{data frame}

\item{x}{numeric column of data}

\item{y}{numeric column of data}

\item{group}{character or factor column of data defining individual lines}

\item{line_colour}{line colour}

\item{line_width}{line width}

\item{xlab}{x-axis title}

\item{ylab}{y-axis title}

\item{title}{plot title}

\item{subtitle}{plot subtitle}

\item{caption}{plot caption}
}
\value{
An object of class ggplot
}
\description{
Produces a simple line chart in ggplot2 which can be customised for publication
}
\details{
The line_chart() function can produce multiple lines if a grouping variable is supplied.
A specific colour can be specified for single lines.
Data should be in long format.
}
\examples{
plot_line_chart(data, x, y)
}
