% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{read_sample_birth_data}
\alias{read_sample_birth_data}
\title{Read Sample Birth Data}
\usage{
read_sample_birth_data()
}
\value{
Data frame with test birth data
}
\description{
Loads the dataframe with test birth data
}
\note{
The Test data in the package is fake and some links between variables will not
represent real relationships.
}
\examples{
test_data_birth <- read_sample_birth_data()

}
