% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{read_sample_death_data}
\alias{read_sample_death_data}
\title{Read Sample Death Data}
\usage{
read_sample_death_data()
}
\value{
A data frame containing the sample death data.
}
\description{
Reads the sample data file created_death_data.csv from the package's extdata directory.
}
