% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{read_sample_divorce_data}
\alias{read_sample_divorce_data}
\title{Read Sample Divorce Data}
\usage{
read_sample_divorce_data()
}
\value{
A data frame containing the sample divorce data.
}
\description{
Reads the sample data file created_divorce_data.csv from the package's extdata directory.
}
