% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{read_sample_marriage_data}
\alias{read_sample_marriage_data}
\title{Read Sample Marriage Data}
\usage{
read_sample_marriage_data()
}
\value{
A data frame containing the sample marriage data.
}
\description{
Reads the sample data file created_marriage_data.csv from the package's extdata directory.
}
