% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_3_1.R
\name{create_t3.1}
\alias{create_t3.1}
\title{Create Table 3.1}
\usage{
create_t3.1(
  bth_data,
  dth_data,
  bth_yr_var,
  dth_yr_var,
  tablename = "Table_4_3",
  output_path = NULL
)
}
\arguments{
\item{bth_data}{Birth Registration Data}

\item{dth_data}{Death Registration Data}

\item{bth_yr_var}{Date of occurrence variable (Birth Data)}

\item{dth_yr_var}{Date of occurrence variable (Death Data)}

\item{tablename}{Name for csv output use _ instead of . for names}

\item{output_path}{The path to export the generated table}
}
\value{
Data frame with tabulated results (if an output_path is given, it will export a .csv)
}
\description{
Table 3.1 Number of vital events by registration timeliness, year
}
\examples{
t3.1 <- create_t3.1(bth_data = bth_data, dth_data = dth_data, bth_yr_var = dobyr, dth_yr_var = dodyr)

}
