% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_4_2.R
\name{create_t4.2}
\alias{create_t4.2}
\title{Creates Table 4.2}
\usage{
create_t4.2(
  data,
  date_var,
  data_year = 2022,
  tablename = "Table_4_2",
  output_path = NULL
)
}
\arguments{
\item{data}{data frame being used}

\item{date_var}{name of date variable}

\item{data_year}{year the data is for}

\item{tablename}{name of the table being saved as a csv file}

\item{output_path}{The path to export the generated table}
}
\value{
data frame of tabulated counts
}
\description{
Live births by place of occurrence and sex of newborn, year

Variables required:
\itemize{
  \item Date of occurrence (birth1a)
  \item Place of occurrence (birth1c)
  \item Type of place of occurrence (hospital, home, etc.) (birth1i)
  \item Sex (birth2a)
}
}
\examples{
t4.2 <- create_t4.2(birth_data, birth1a, data_year = 2022)
}
