% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_4_3.R
\name{create_t4.3}
\alias{create_t4.3}
\title{Create Table 4.3}
\usage{
create_t4.3(data, date_var, data_year = 2022, tablename = "Table_4_3")
}
\arguments{
\item{data}{Birth registration data}

\item{date_var}{Date of occurrence variable}

\item{data_year}{Chosen year}

\item{tablename}{Name for csv output use _ instead of . for names}
}
\value{
Data frame with tabulated results
}
\description{
Live births by place of occurrence and place of usual residence of mother, year

Variables required:
\itemize{
  \item Date of occurrence (birth1a)
  \item Place of occurrence (birth1c)
  \item Place of usual residence - Mother (birth3l)
}
}
\note{
The Vital Statistics \href{https://www.vitalstrategies.org/resources/production-of-a-vital-statistics-report/}{report resource kit}
includes guidance and templates to help complete your CRVS report.
}
\examples{
t4.3 <- create_t4.3(data, date_var = year, data_year = 2022, tablename = "Table_4_3")
}
