% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_5_1.R
\name{create_t5.1}
\alias{create_t5.1}
\title{Calculates Table 5.1 Deaths summary table}
\usage{
create_t5.1(data, num_yrs = 5, tablename = "Table_5_1")
}
\arguments{
\item{data}{deaths data frame}

\item{num_yrs}{number of years to report on}

\item{tablename}{name for csv output use _ instead of . for names}
}
\value{
data frame with tabulated results
}
\description{
Calculates Table 5.1 Deaths summary table
}
\examples{
t5.1 <- create_t5.1(dth_data, tablename = "Table_5_1")

}
