% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_5_2.R
\name{create_t5.2}
\alias{create_t5.2}
\title{Creates Tables 5.2}
\usage{
create_t5.2(
  data,
  date_var = "dodyr",
  data_year = NA,
  tablename = "Table_5_2",
  output_path = NULL
)
}
\arguments{
\item{data}{Death records data frame.}

\item{date_var}{Event year column of the dataframe. Defaults to "dodyr".}

\item{data_year}{The year to report on. Defaults to the last in the data.}

\item{tablename}{Name of the table to be saved as a csv file. Optional.}

\item{output_path}{The path to export the generated csv table. Optional.}
}
\value{
A data frame of tabulated results
}
\description{
This function generates Table 5.2, which shows deaths by place of usual residence
and the sex of the deceased.
}
\details{
Variables needed per record:

- `dodyr`: Year of death.

- `death2o`: Region of usual residence.

- `death2c`: Sex of the deceased.
}
\examples{
death_data <- construct_sample_death_data()
death_data <- construct_year(death_data, date_col = "death1a", year_col = "dodyr")
table_5_2 <- create_t5.2(death_data, date_var="dodyr", data_year = 2022)
}
